+++ src/libstddjb/buffer_init.c
@@ -1,11 +1,12 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <errno.h>
 #include <skalibs/cbuffer.h>
 #include <skalibs/functypes.h>
 #include <skalibs/buffer.h>
 
-int buffer_init (buffer *b, iovfunc_t_ref op, int fd, char *s, unsigned int len)
+int buffer_init (buffer *b, iovfunc_t_ref op, int fd, char *s, size_t len)
 {
   if (!cbuffer_init(&b->c, s, len)) return 0 ;
   b->fd = fd ;
