+++ src/libstddjb/buffer_putflush.c
@@ -1,10 +1,11 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/buffer.h>
 
-int buffer_putflush (buffer *b, char const *s, unsigned int len)
+ssize_t buffer_putflush (buffer *b, char const *s, size_t len)
 {
-  int r = buffer_put(b, s, len) ;
+  ssize_t r = buffer_put(b, s, len) ;
   if (r < 0) return -1 ;
   if (!buffer_flush(b)) return -1 ;
   return r ;
