+++ src/libstddjb/pathexec_r_name.c
@@ -1,12 +1,13 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/bytestr.h>
 #include <skalibs/env.h>
 #include <skalibs/djbunix.h>
 
-void pathexec_r_name (char const *file, char const *const *argv, char const *const *envp, unsigned int envlen, char const *modifs, unsigned int modiflen)
+void pathexec_r_name (char const *file, char const *const *argv, char const *const *envp, size_t envlen, char const *modifs, size_t modiflen)
 {
-  unsigned int n = envlen + 1 + byte_count(modifs, modiflen, '\0') ;
+  size_t n = envlen + 1 + byte_count(modifs, modiflen, '\0') ;
   char const *v[n] ;
   if (env_merge(v, n, envp, envlen, modifs, modiflen))
     pathexec_run(file, argv, v) ;
