+++ src/libstddjb/string_quote_nodelim.c
@@ -1,9 +1,10 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/stralloc.h>
 #include <skalibs/skamisc.h>
 
-int string_quote_nodelim (stralloc *sa, char const *s, unsigned int len)
+int string_quote_nodelim (stralloc *sa, char const *s, size_t len)
 {
   return string_quote_nodelim_mustquote(sa, s, len, "\"", 1) ;
 }
