+++ src/libunixonacid/openwritenclose.c
@@ -2,10 +2,11 @@
 
 /* MT-unsafe */
 
+#include <sys/types.h>
 #include <skalibs/skamisc.h>
 #include <skalibs/unix-transactional.h>
 
-int openwritenclose (char const *fn, char const *s, unsigned int len)
+int openwritenclose (char const *fn, char const *s, size_t len)
 {
   return openwritenclose_tmp(fn, s, len, &satmp) ;
 }
