+++ src/libunixonacid/unixmessage_receiver_init.c
@@ -1,11 +1,12 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <errno.h>
 #include <skalibs/cbuffer.h>
 #include <skalibs/stralloc.h>
 #include <skalibs/unixmessage.h>
 
-int unixmessage_receiver_init (unixmessage_receiver_t *b, int fd, char *mainbuf, unsigned int mainlen, char *auxbuf, unsigned int auxlen)
+int unixmessage_receiver_init (unixmessage_receiver_t *b, int fd, char *mainbuf, size_t mainlen, char *auxbuf, size_t auxlen)
 {
   if (!cbuffer_init(&b->mainb, mainbuf, mainlen)
    || !cbuffer_init(&b->auxb, auxbuf, auxlen)) return 0 ;
